# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the SplitBill API documentation generated by Swag.",
        "title": "SplitBill API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Kajal Rana",
            "email": "kajalrana@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/{user_name}": {
            "get": {
                "description": "Retrieve user details",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a user by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Name",
                        "name": "user_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}